Sync gethostname() with FreeBSD:
* Change gethostname() to set errno to ENAMETOOLONG instead of ENOMEM
when the buffer is not long enough to hold the current host name.
POSIX does not standardize error returns for gethostname(), so it
doesn't matter which one we use, but ENAMETOOLONG is at least a little
more intuitive.

* Update prototype to match SuS (int->size_t).

+++ include/unistd.h
@@ -153,7 +153,7 @@ int	 getdomainname(char *, int);
 int	 getdtablesize(void);
 int	 getgrouplist(const char *, gid_t, gid_t *, int *);
 long	 gethostid(void);
-int	 gethostname(char *, int);
+int	 gethostname(char *, size_t);
 int	 getlogin_r(char *, int);
 mode_t	 getmode(const void *, mode_t);
 int	 getpagesize(void) __pure2;
